feat: add support for lerna mono-repos
If the project is a lerna mono-repo it will add an extra question about
the packages the commit affected
