+++ .gitignore
@@ -1,3 +1,4 @@
+*.log
 *.swo
 *.swp
 .DS_Store
