--- .jscsrc
@@ -1,8 +1,0 @@
-{
- "preset": "airbnb",
-
- "requireTrailingComma": false,
- "maximumLineLength": 200,
- "disallowMultipleVarDecl": false,
- "safeContextKeyword": ["_this"]
-}
