--- .jshintrc
@@ -1,25 +1,0 @@
-{
-  "bitwise": true,
-  "browser": true,
-  "browserify": true,
-  "devel": true,
-  "eqnull": true,
-  "expr": true,
-  "esversion": 6,
-  "immed": true,
-  "indent": 2,
-  "laxbreak": true,
-  "loopfunc": true,
-  "maxlen": 200,
-  "mocha": true,
-  "newcap": true,
-  "noarg": true,
-  "noempty": true,
-  "nonew": true,
-  "strict": true,
-  "sub": true,
-  "trailing": true,
-  "undef": true,
-  "unused": true,
-  "globals":{}
-}
