+++ .nvmrc
@@ -1,1 +1,1 @@
-6
+8
