+++ README.md
@@ -2,7 +2,6 @@
 
 # mol-conventional-changelog
 
-
 ## Setup
 
 Install commitizen
@@ -94,6 +93,10 @@ The subject contains succinct description of the change:
 
 Just as in the **subject**, use the imperative, present tense: "change" not "changed" nor "changes".
 The body should include the motivation for the change and contrast this with previous behavior.
+
+#### Affects [only on [lerna](https://lernajs.io/) environments]
+
+Select the packages the commit affected.
 
 ### Breaking Changes
